backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.11.7,<3
pydantic-settings>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
websockets>=12.0
python-multipart>=0.0.9
runpod-flash
python-dotenv>=1.0.0
openai>=2.20.0
supabase >=2.28.0
google-cloud-storage>=2.14.0